Move .permissions.php and .listAll.php
